Fix tld, and iprange when writing to greencheck tbl
We parse out the tld, and check for a deleted ip ragnge
when logging now.

Add mgmt command for consuming from the queue

We need to switch off timezone support if we want
to touch the greencheck table, as it uses timestamps

Add migrations back in for tests.

At least this way we know our tests work against a
realistic representation of the prod database

Add tld, and remove unneeded UnixDateTimeField

Split out parsing from logging for sitechecks