Domain claiming via LinkedDomains v1.
This commit introduces a new mechanism, feature flagged at present, for
linking a domain to a provider using carbon txt.

Hosting provider users may link domains to their provider, either via
hosting a carbon.txt file, or using a delegation mechamism to point to
an existing provider which has been linked already.

When a greencheck is done for a linked domain, it will show as "green"
IF the LinkedDomain has been approved, AND the associated provider is
shown on the site and not archvied

- LinkedDomains must be verified manually be the GWF team for now
- LinkedDomains may be "primary" (representing the provider itself), or
  "secondary" (represnting a customer)
- This replaces the old carbon_txt_url field - all existing
  carbon_txt_urls have been migrated to LinkedDomains, and any providers
  which subsequently create a LinkedDomain will have the carbon txt
  badge in the directory.