Add carbon.txt badge in directory
This commit adds a new carbon_txt_url field to hosting providers, which
can be populated in the admin console, and a "carbon txt" badge to be
shown on directory listings for those providers which have a carbon_txt
url set. This allows us to highlight the presence of carbon.txt in the
directory for providers that are using it, while side-stepping some of
the issues pending resolution with domain verification for now.