Generate and run all pending migrations
We've had a bunch of migrations automatically generated with
`makemigrations` for a while that we've never run, as we were worried
they might be unintentionally destructive of data: due to the fact that
MariaDB enums rely on the order in which values are defined, changing
the possible values of an enum might change the actual stored values, if
the new value set doesn't maintain the same order.

I've tested this and verified that no data is changed (verifying that
the order of enum definitions remain the same, and that a `select ENUM_FIELD
count(id) group by ENUM_FIELD` returns the same totals before and after
the migration).

On deployment, it'd be worth:

- **taking a database backup before running the migrations**
- **running the same tests**: `select type, count(id) from greencheck group
  by type order by count(id) desc;` and `select type, count(id) from greencheck group by type order by count(id) desc;` before and after running the migration, and ensuring that the totals stay the same.

  Once this is run, the django database schema, django model
  definitions, and database schema should all be in sync, and `makemigrations` will not create masses of spurious changes everytime we run it.