Implement caching of carbon.txt checks,
Ensuring that the greenweb check api endpoint remains responsive, and we
don't inadvertently DOS attack frequently checked domains. This also
ensures that the relevant cache date is returned in grey greencheck
results, and that the domain cache respects nocache behaviour.

An exception is the image endpoint, where nocache is ignored at the
carbon.txt level. This is because we've historically provided nocache
image urls in the embed code, and therefore the uncached image endpoint
is hit especially hard.