The clone_db script is meant to be used in staging, in order to produce a complete, safe, copy of the production DB and assets for testing purposes.
The original motivation for producing this script was to avoid the
possibility of data loss if URLs for production assets were copied over
to staging in the db dump (the theory being that a deletion on staging
would then delete the production asset), but this in fact isn't the case
- the URLs / keys of assets are the same on staging / production, but
  they use different buckets, therefore, in the worst case, the asset
  wouldn't be available on staging, and production would be unaffected.

It still seems useful to have a way of producing a complete
copy of production on staging. However, I'm a little concerned that we've now *introduced* a
vector for data loss, rather than eliminating one - running the clone_db
script **in production** has the potential to wipe out production
data if incorrectly configured.

I've attempted to document the behaviour clearly and comprehensively,
and include plenty of manual confirmation to prevent this from
happening, but perhaps there's a conversation to be had about whether
the benefits of having this script at all (uploaded files will be
accessible in staging after restoring the db from a dump) outweigh the
risks (running the script in production erroneously could lead to data
loss).